# Applications/QuizManager/QuizManager.Core/Models/Answer.cs
namespace QuizManager.Core.Models
{
    public class Answer
    {
        public int? Id { get; set; }
        public string Text { get; set; }
        public bool IsCorrect { get; set; }
        public int QuestionId { get; set; }
    }
}